+++ android/fakewallet/src/androidTest/java/com/solana/mobilewalletadapter/walletlib/scenario/DevicePowerConfigProvider.kt
@@ -1,10 +1,11 @@
-package com.solana.mobilewalletadapter.walletlib.provider
+package com.solana.mobilewalletadapter.walletlib.scenario
 
 import android.content.Context
 
 var TestScopeLowPowerMode: Boolean = false
 
-class DevicePowerConfigProvider(context: Context) : PowerConfigProvider {
+class DevicePowerConfigProvider(context: Context) :
+    PowerConfigProvider {
     override fun isLowPowerMode(): Boolean {
         return TestScopeLowPowerMode
     }
