+++ android/fakewallet/src/main/java/com/solana/mobilewalletadapter/fakewallet/MobileWalletAdapterActivity.kt
@@ -27,24 +27,17 @@ class MobileWalletAdapterActivity : AppCompatActivity() {
                 viewModel.mobileWalletAdapterServiceEvents.collect { request ->
                     if (request is MobileWalletAdapterServiceRequest.SessionTerminated) {
                         finish()
+                    } else if (request is MobileWalletAdapterServiceRequest.LowPowerNoConnection) {
+                        // should use dialog fragment, etc. but this is a quick demo
+                        AlertDialog.Builder(this@MobileWalletAdapterActivity)
+                            .setTitle(R.string.low_power_mode_warning_title)
+                            .setMessage(R.string.str_low_power_mode_warning_dsc)
+                            .setPositiveButton(android.R.string.ok) { _, _ ->
+                                Log.w(TAG, "Connection failed due to device low power mode, returning to dapp.")
+                                finish()
+                            }
+                            .show()
                     }
-                }
-            }
-        }
-
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.CREATED) {
-                viewModel.lowPower.collect {
-                    // should use dialog fragment, etc. but this is a quick demo
-                    AlertDialog.Builder(this@MobileWalletAdapterActivity)
-                        .setTitle(R.string.low_power_mode_warning_title)
-                        .setMessage(R.string.str_low_power_mode_warning_dsc)
-                        .setPositiveButton(R.string.label_return_to_dapp) { _, _ ->
-                            Log.w(TAG, "Connection failed due to device low power mode, returning to dapp.")
-                            finish()
-                        }
-                        .setNeutralButton(R.string.label_close_dialog) { d, _ -> d.dismiss() }
-                        .show()
                 }
             }
         }
