--- android/fakewallet/src/main/res/values/constants.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <integer name="no_connection_timeout_millis">3000</integer>
-</resources>
