--- android/walletlib/src/main/java/com/solana/mobilewalletadapter/walletlib/provider/PowerConfigProvider.java
@@ -1,5 +1,0 @@
-package com.solana.mobilewalletadapter.walletlib.provider;
-
-public interface PowerConfigProvider {
-    boolean isLowPowerMode();
-}
