+++ android/walletlib/src/main/java/com/solana/mobilewalletadapter/walletlib/scenario/PowerConfigProvider.java
@@ -1,0 +1,5 @@
+package com.solana.mobilewalletadapter.walletlib.scenario;
+
+interface PowerConfigProvider {
+    boolean isLowPowerMode();
+}
