--- android/walletlib/src/test/java/android/util/Log.java
@@ -1,29 +1,0 @@
-package android.util;
-
-/**
- * Mocks Android's Log class (android.util.Log) so that we can test code that uses the Log
- * methods without needed to manually mock or inject a logging abstraction for every test.
- *
- * This mock is very basic, it does the minimum amount of work to provide a functional test env.
- */
-public class Log {
-    public static int d(String tag, String msg) {
-        System.out.println("DEBUG: " + tag + ": " + msg);
-        return 0;
-    }
-
-    public static int i(String tag, String msg) {
-        System.out.println("INFO: " + tag + ": " + msg);
-        return 0;
-    }
-
-    public static int w(String tag, String msg) {
-        System.out.println("WARN: " + tag + ": " + msg);
-        return 0;
-    }
-
-    public static int e(String tag, String msg) {
-        System.out.println("ERROR: " + tag + ": " + msg);
-        return 0;
-    }
-}
